+++ packages/aio-commerce-lib-app/tsdown.config.ts
@@ -18,14 +18,17 @@ export default mergeConfig(baseConfig, {
   copy: [
     {
       from: "./source/commands/generate/actions/templates",
-      to: "./dist/cjs/commands/generate/actions/templates",
+      to: "./dist/cjs/commands/generate/actions",
     },
     {
       from: "./source/commands/generate/actions/templates",
-      to: "./dist/es/commands/generate/actions/templates",
+      to: "./dist/es/commands/generate/actions",
     },
   ],
 
-  // This package is private and needs to be bundled as a no-external dependency.
-  noExternal: ["@aio-commerce-sdk/scripting-utils"],
+  // These packages are private and need to be bundled as no-external dependencies.
+  noExternal: [
+    "@aio-commerce-sdk/scripting-utils",
+    "@aio-commerce-sdk/common-utils",
+  ],
 });