+++ packages/aio-commerce-lib-auth/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @adobe/aio-commerce-lib-auth
 
+## 0.8.0
+
+### Minor Changes
+
+- [#217](https://github.com/adobe/aio-commerce-sdk/pull/217) [`6063a97`](https://github.com/adobe/aio-commerce-sdk/commit/6063a97299c9b44f663314420b59cee7998b2f89) Thanks [@iivvaannxx](https://github.com/iivvaannxx)! - Introduced CLI utility `sync-ims-credentials` to sync the workspace credentials of `.aio` into the `.env` file environment variables expected by the library
+
+## 0.7.0
+
+### Minor Changes
+
+- [#224](https://github.com/adobe/aio-commerce-sdk/pull/224) [`ac46985`](https://github.com/adobe/aio-commerce-sdk/commit/ac46985186961c65eddc6be3200dbe1e00369055) Thanks [@iivvaannxx](https://github.com/iivvaannxx)! - Implement token forwarding utilities
+
+### Patch Changes
+
+- Updated dependencies [[`f2b6f34`](https://github.com/adobe/aio-commerce-sdk/commit/f2b6f34709a5adcca93b852b5855cb6b1c852312), [`f2b6f34`](https://github.com/adobe/aio-commerce-sdk/commit/f2b6f34709a5adcca93b852b5855cb6b1c852312)]:
+  - @adobe/aio-commerce-lib-core@0.6.0
+
 ## 0.6.2
 
 ### Patch Changes